Finalization of the work
In the portfolio cards, I reduced the text on the mobile and tablet versions and changed the indents. For the list of social networks in the mobile menu, I also adapted the text. In the modal window, I made the text "I agree with the newsletter and accept the terms of the contract" as one block and adapted the size of the text for different viewport sizes